+++ packages/web-components/src/components/global-header/components/global-header/src/components/HybridIpaasHeader/HybridIpaasHeader.ts
@@ -10,11 +10,12 @@
 
 import { css, LitElement, html, unsafeCSS } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
-import { settings } from '@carbon-labs/utilities/es/settings/index.js';
+import { settings } from '@carbon-labs/utilities';
 import {
   CUSTOM_EVENT_NAME,
   CUSTOM_EVENT_DETAIL_REFRESH_OPTIONS,
   INITIAL_AUTOMATION_HEADER_PROPS,
+  SOLIS_CDN_HOSTNAMES,
 } from '../../constant';
 import {
   GlobalActionConfig,
@@ -44,6 +45,7 @@ export class HybridIpaasHeader extends LitElement {
   @property({ type: Object }) fetchHeaders = {};
   @property({ type: Boolean }) solisSidekickEnabled = false;
   @property({ type: Boolean }) solisSwitcherEnabled = false;
+  @property({ type: String }) solisEnvironment = 'local';
   @property({ type: String }) basePath = '';
   @property({ type: String }) displayName = '';
   @property({ type: String }) userEmail = '';
@@ -52,9 +54,15 @@ export class HybridIpaasHeader extends LitElement {
   @property({ type: Function }) aiCallback: (() => void) | undefined;
   @property({ type: String }) aiCallbackEvent = '';
   @property({ type: Function }) logoutCallback: (() => void) | undefined;
+  @property({ type: String }) logoutCallbackEvent = '';
   @property({ type: Function }) notificationOpenCallback:
     | (() => void)
     | undefined;
+  @property({ type: String }) notificationOpenCallbackEvent = '';
+  @property({ type: Function }) searchCallback: (() => void) | undefined;
+  @property({ type: String }) searchCallbackEvent = '';
+  @property({ type: Function }) searchSubmitCallback: (() => void) | undefined;
+  @property({ type: String }) searchSubmitCallbackEvent = '';
   @property({ type: Boolean }) hasNewNotifications = false;
   @property({ type: Object }) searchConfigs: SearchConfigs | null = null;
   @property({ type: Array })
@@ -140,6 +148,14 @@ export class HybridIpaasHeader extends LitElement {
 
     if (this.logoutCallback) {
       footerLink.onClickHandler = this.logoutCallback;
+    } else if (this.logoutCallbackEvent) {
+      footerLink.onClickHandler = () => {
+        const event = new CustomEvent(this.logoutCallbackEvent, {
+          bubbles: true,
+          cancelable: true,
+        });
+        this.dispatchEvent(event);
+      };
     } else {
       footerLink.href = `${this.basePath}/logout`;
     }
@@ -147,33 +163,33 @@ export class HybridIpaasHeader extends LitElement {
     return footerLink;
   }
 
-  private initSidekickOptions() {
+  private initSidekickOptions(env = 'local') {
     return {
       isEnabled: this.solisSidekickEnabled,
-      scriptUrl:
-        'https://cdn.dev.saas.ibm.com/solis_ui/v1/sidekick/solis-sidekick.es.js',
+      scriptUrl: SOLIS_CDN_HOSTNAMES[env] + '/sidekick/solis-sidekick.es.js',
       insights_enabled: true,
       reports_enabled: true,
       chat_enabled: false,
       tell_me_more_enabled: false,
     };
   }
-  private initSolisOptions(forSidekick = false) {
+  private initSolisOptions(env = 'local', forSidekick = false) {
     if (forSidekick) {
       return {
         isEnabled: false,
-        is_prod: false,
-        cdn_hostname: 'https://cdn.dev.saas.ibm.com/solis_ui/v1',
-        deployment_environment: solisDeploymentEnvironment['local'],
+        is_prod: env === 'prod',
+        cdn_hostname: SOLIS_CDN_HOSTNAMES[env],
+        deployment_environment: solisDeploymentEnvironment[env] || 'local',
+        product_id: 'ipaas',
       };
     } else {
       return {
         isEnabled: true,
-        scriptUrl:
-          'https://cdn.dev.saas.ibm.com/solis_ui/v1/switcher/solis-switcher.es.js',
-        is_prod: false,
-        cdn_hostname: 'https://cdn.dev.saas.ibm.com/solis_ui/v1',
-        deployment_environment: solisDeploymentEnvironment['local'],
+        scriptUrl: SOLIS_CDN_HOSTNAMES[env] + '/switcher/solis-switcher.es.js',
+        is_prod: env === 'prod',
+        cdn_hostname: SOLIS_CDN_HOSTNAMES[env],
+        deployment_environment: solisDeploymentEnvironment[env] || 'local',
+        product_id: 'ipaas',
       };
     }
   }
@@ -249,12 +265,47 @@ export class HybridIpaasHeader extends LitElement {
         onClick: this.notificationOpenCallback,
       };
     }
+    if (this.notificationOpenCallbackEvent) {
+      updatedOptions.notificationConfigs = {
+        onClick: () => {
+          const event = new CustomEvent(this.notificationOpenCallbackEvent, {
+            bubbles: true,
+            cancelable: true,
+          });
+          this.dispatchEvent(event);
+        },
+      };
+    }
 
     if (this.searchConfigs) {
       updatedOptions.searchConfigs = {
         placeholder: this.searchConfigs?.placeholder ?? 'Search',
         callback: this.searchConfigs?.callback,
         submitCallback: this.searchConfigs?.submitCallback,
+      };
+    }
+
+    if (this.searchCallbackEvent) {
+      updatedOptions.searchConfigs = {
+        callback: () => {
+          const event = new CustomEvent(this.searchCallbackEvent, {
+            bubbles: true,
+            cancelable: true,
+          });
+          this.dispatchEvent(event);
+        },
+      };
+    }
+
+    if (this.searchSubmitCallbackEvent) {
+      updatedOptions.searchConfigs = {
+        submitCallback: () => {
+          const event = new CustomEvent(this.searchSubmitCallbackEvent, {
+            bubbles: true,
+            cancelable: true,
+          });
+          this.dispatchEvent(event);
+        },
       };
     }
 
@@ -276,11 +327,16 @@ export class HybridIpaasHeader extends LitElement {
     }
 
     if (this.solisSidekickEnabled) {
-      updatedOptions.solisConfig = this.initSolisOptions(true);
-      updatedOptions.sidekickConfig = this.initSidekickOptions();
+      updatedOptions.solisConfig = this.initSolisOptions(
+        this.solisEnvironment,
+        true
+      );
+      updatedOptions.sidekickConfig = this.initSidekickOptions(
+        this.solisEnvironment
+      );
     }
     if (this.solisSwitcherEnabled) {
-      updatedOptions.solisConfig = this.initSolisOptions();
+      updatedOptions.solisConfig = this.initSolisOptions(this.solisEnvironment);
     }
 
     if (