+++ packages/web-components/src/components/global-header/components/global-header/src/constant.ts
@@ -68,3 +68,10 @@ export const DEFAULT_NO_AUTH_HEADER_LINKS = [
     arialLabel: 'Log in',
   },
 ];
+
+export const SOLIS_CDN_HOSTNAMES = {
+  local: 'https://cdn.dev.saas.ibm.com',
+  dev: 'https://cdn.dev.saas.ibm.com',
+  stage: 'https://cdn.test.saas.ibm.com',
+  prod: 'https://cdn.saas.ibm.com',
+};