+++ packages/web-components/src/components/global-header/components/global-header/src/globals/__tests__/loadSidekickScript.test.ts
@@ -57,6 +57,7 @@ const propsWithSidekickConfig: HeaderProps = {
     is_prod: false,
     cdn_hostname: 'https://cdn.dev.saas.ibm.com/solis_ui/v1',
     deployment_environment: solisDeploymentEnvironment['local'],
+    product_id: 'my_product',
   },
   switcherConfigs: [
     {
@@ -374,6 +375,10 @@ describe('loadSidekickScript function', () => {
     );
     expect((window as any)._solis.deployment_environment).to.equal(
       propsWithSidekickConfig.solisConfig?.deployment_environment
+    );
+
+    expect((window as any)._solis.product_id).to.equal(
+      propsWithSidekickConfig.solisConfig?.product_id
     );
 
     expect((window as any)._solis.sidekick.correlation_id).to.equal(