+++ packages/web-components/src/components/global-header/components/global-header/src/globals/utils.ts
@@ -124,6 +124,7 @@ export function getAssistMeUrl(environment: string) {
 }
 
 export const trackEvent = (eventName: string, props: EventProps) => {
+  // console.log('trackEvent', eventName, props) // For testing purposes
   const analytics =
     /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
     (window as any)?.bluemixAnalytics;