+++ packages/web-components/src/components/global-header/components/global-header/src/types/Header.types.ts
@@ -56,7 +56,7 @@ export interface ChatBotConfigs {
 
 export interface SearchConfigs {
   placeholder?: string;
-  callback: (value: string) => void;
+  callback?: (value: string) => void;
   submitCallback?: (value: string) => void;
 }
 
@@ -296,18 +296,45 @@ export interface HeaderProps {
   sidekickConfig?: SidekickConfig;
   solisConfig?: SolisConfig;
 }
-export enum solisDeploymentEnvironment {
-  'local',
-  'dev',
-  'stage',
-  'prod',
+
+export interface ReactWrapperProps
+  extends Omit<HeaderProps, 'chatBotConfigs' | 'notificationConfigs'> {
+  productKey: string;
+  productName?: null;
+  fetchHeaders?: Record<string, string>;
+  solisSidekickEnabled?: boolean;
+  solisSwitcherEnabled?: boolean;
+  solisEnvironment?: string;
+  basePath?: string;
+  displayName?: string;
+  userEmail?: string;
+  productVersion?: null;
+  assistMeKey?: string;
+  hasNewNotifications?: boolean;
+  capabilityProfileFooterLinks?: ProfileFooterLinks[];
+  capabilityGlobalActions?: GlobalActionConfig[];
+  searchConfigs?: Omit<SearchConfigs, 'callback' | 'submitCallback'>;
+  aiCallback?: () => void | undefined;
+  notificationOpenCallback?: () => void | undefined;
+  logoutCallback?: () => void | undefined;
+  searchCallback?: (value: string) => void | undefined;
+  searchSubmitCallback?: (value: string) => void | undefined;
 }
+
+export enum solisDeploymentEnvironment {
+  local = 'local',
+  dev = 'dev',
+  stage = 'stage',
+  prod = 'prod',
+}
+
 export interface SolisConfig {
   isEnabled: boolean;
   scriptUrl?: string;
   is_prod?: boolean;
   cdn_hostname: string;
   deployment_environment: solisDeploymentEnvironment;
+  product_id: string;
 }
 
 export interface NavigationItem {
@@ -336,8 +363,9 @@ export type SidekickInfo = {
 interface solisWindowConfig {
   is_prod?: boolean;
   cdn_hostname: string;
-  deployment_environment: 'local' | 'dev' | 'stage' | 'prod';
-  sidekick: SidekickInfo;
+  deployment_environment: solisDeploymentEnvironment;
+  sidekick?: SidekickInfo;
+  product_id: string;
 }
 
 declare global {